Adding code for task FT-17
1. Importing all constasts and library
2. Adding to the refs properties that were in the function for opening modal
3. Deleting repeating refs in first function that open modal
4. Adding code for filling modal markup by full info about film
5. Adding import the modal_film.js file into index.js file
6. Deleting tag script that connected the modal_film.js file from index.html

Also I see, that tag script for connection index.js file are adding twice: in the tag  head and in the end of body tag.  But I didn't delete any of them. I think it will be better to delete script tag in head as Repeta said but if you don't agree with me, delete another one.